+++ android/viro_bridge/src/main/java/com/viromedia/bridge/component/node/control/VideoSurfaceManager.java
@@ -92,7 +92,7 @@ public class VideoSurfaceManager extends NodeManager<VideoSurface> {
     public void receiveCommand(VideoSurface video, int commandType, @Nullable ReadableArray args) {
         switch (commandType) {
             case ViroCommands.SEEK_TO_TIME_INDEX:
-                video.seekToTime((int) args.getDouble(0));
+                video.seekToTime((float) args.getDouble(0));
                 break;
             default:
                 throw new IllegalArgumentException("Unsupported command " + commandType
