+++ ios/ViroReact/Views/VRTVideoSurface.mm
@@ -34,7 +34,7 @@
     return self;
 }
 
-- (void)seekToTime:(NSInteger)time {
+- (void)seekToTime:(float)time {
     if (_videoTexture) {
         _videoTexture->seekToTime(time);
     }
@@ -178,7 +178,7 @@
     }
 }
 
-- (void)videoDidUpdateTime:(int)currentTimeInSeconds totalTimeInSeconds:(int)totalTime{
+- (void)videoDidUpdateTime:(float)currentTimeInSeconds totalTimeInSeconds:(float)totalTime{
     if (self.onUpdateTimeViro) {
         self.onUpdateTimeViro(@{@"currentTime": @(currentTimeInSeconds),
                                 @"totalTime": @(totalTime)});
